requirements.txt
appnope==0.1.2
autopep8==1.5.6
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
decorator==5.0.7
idna==2.10
install==1.3.4
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
memory-profiler==0.58.0
numpy==1.21.1
pandas==1.3.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.8.1
pymongo==3.12.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
webdriver-manager==3.4.2
